fix: register service at the first place
